cleaned up App.js. Moved logic and api call to Game.js. Removed extra pseudo code. Game and start comps are working as they should
